refactor: extract platform specific code
