+++ packages/webview/src/browser/index.ts
@@ -2,6 +2,7 @@ import { Provider, Injectable } from '@opensumi/di';
 import { BrowserModule } from '@opensumi/ide-core-browser';
 
 import { WebviewModuleContribution } from './contribution';
+import { ElectronWebviewModuleContribution } from './electron.contribution';
 import { IWebviewService } from './types';
 import { WebviewServiceImpl } from './webview.service';
 export * from './types';
@@ -16,4 +17,5 @@ export class WebviewModule extends BrowserModule {
     },
     WebviewModuleContribution,
   ];
+  electronProviders: Provider[] = [ElectronWebviewModuleContribution];
 }
